Stopped git tracking the nef files
In the python/ dir, there is a dir, nef_data, which stores data needed
for the quick creation of NEF ensembles. These really shouldn't be
tracked, as they are automatically generated and there will always
be a growing number of them. So we no longer track them. Only tracked
them in the first place as a mistake.