Refactored the control script.
Refactored the control script so that it is much more readable. Took
most of the functionality out of the script itself and put it into a
module start_utils.py, so that this functionality is more reusable and
we can easily write alternative control scripts. Also, renamed the
control script to test_script.py inline with python naming conventions.