Fixed global variable scheme for sym definition
Symbols denoting which relations are to be used were defined in
SymbolDefinitions.py as global variables. Fixed this so that we
get them via calling functions since global variables are bad.
Also allows us to choose which symbol definitions we want to use.

Also, added a command line arg to easily control whether we are using
bidirectional or unidirectional relations. Just add -b to get
bidirectional. Unidirectional is default.

Finally, changed name of SymbolDefinitions.py to adhere to python naming
conventions.