Lots of changes, especially for unitary & id vecs
* made unitary and id vectors work. can use them independently of one
another, if unitary is on but id isn't, the the relations will be
unitary (they're the only things that would make a difference)

* moved the data collection into Bootstrapper class in bootstrap.py

* added data collection in assoc_memory.py to get raw stats on HRR
results so we can see what the results should be in principle.

* added a header into the output files showing the configuration of
the run

* made random number generation work right. Before, it was using
the same RNG to generate both the vectors and a test plan. This
would cause test plans under different vector-generation
conditions to get out of synch, since different numbers of calls
to the RNG would be required. Now, each of these stages has its own
RNG.

* got rid of the args, kwargs that used to be in almost every function
didn't use them for anything

* added a bunch of command line args, so you can specify what type
of test, how many trials and runs, and whether to use id vecs/unitary
vecs