Fixed hier testing with assoc mem, extended .pdbrc
* Previously, hier test wouldn't work with standard associative
memory. The assoc memory returns keys instead of vectors, because the
vectors it would return would be exactly the same as a structured
vector, so this saves some searching through the vocabulary. However,
the hier test assumed that if use_HRR = true, then we would be passing
around vectors. Fixed it so it asks the associator whether it wants to
return vectors or keys, and adjusts accordingly.

* Extended .pdbrc with another alias.