Made a new class for doing just WordNet testing
* AssociativeMemoryTester used to be a giant class. However, a lot
of the code in there was specific to testing using WordNet as the
underlying graph, which we don't always want to do if we want to
run quick, more controlled tests.

* So we made a new class WordnetAssociativeMemoryTester which
inherits from AssociativeMemoryTester, and contains all the code
for using WordNet, while AssociativeMemoryTester becomes a class
that can be used to quickly test associative memories.