Added option to the neural alg w/o neurons
The neural algorithm is to basically threshold all the dot products
and multiply the output vectors by these threshold values and sum up
all the resulting vectors. Now have an option to use this same
algorithm without the neural model.

Specify this mode with the -a command line arg. Overrides the -n
command line arg.