Changed for symbols are chosen for the h-test
Now there is a function in symbol_definitions.py called
hierarchical_symbols which returns the symbols to be used as queries
on the hier test. This makes it easier to choose which symbols we
want to use for a given hierarchical test, since we only have to
change one line.