Encapsulates probe functionality. Fixes #2.
* the Probe class in probe.py now records the history of a run.

* each probe can store data from multiple runs

* currently, probes are created be calling probe_by_connection
on the ensemble that we want to probe. internally, the probe creates
an ArrayNode() and connects that to the ensemble passed into
probe_by_connection. The probe gets its data by calling "array" on
that internal node.

* probes have their own plot function, the cleanup network plot
function works by simply calling the plot function of each of its
probes in succession