Lots of changes to class structure.
Generally trying to make everything more modular, and have objects
use one another through well-defined API's. Especially true for how the
different objects interact with files.

The experiment is given an output filename, where it is to print
results. The experiment itself has no say in what this gets called. The
experiment will then create a directory based on this name (using a
slightly different name), and then all the little peices of the
experiment can write their log files to that dir. The result is that
each experiment has a file giving the overall results, and a directory
where logging information is stored.